Move memtable related files into memtable directory
Summary:
Move memtable related files into memtable directory.

Test Plan:
run existing tests.
