+++ db/flush_job.cc
@@ -24,7 +24,6 @@
 #include "db/event_helpers.h"
 #include "db/log_reader.h"
 #include "db/log_writer.h"
-#include "db/memtable.h"
 #include "db/memtable_list.h"
 #include "db/merge_context.h"
 #include "db/version_set.h"
@@ -33,6 +32,7 @@
 #include "monitoring/thread_status_util.h"
 #include "port/likely.h"
 #include "port/port.h"
+#include "db/memtable.h"
 #include "rocksdb/db.h"
 #include "rocksdb/env.h"
 #include "rocksdb/statistics.h"
