+++ db/write_thread.cc
@@ -9,6 +9,7 @@
 #include <thread>
 #include "db/column_family.h"
 #include "port/port.h"
+#include "util/random.h"
 #include "util/sync_point.h"
 
 namespace rocksdb {
