+++ memtable/memtable_allocator.cc
@@ -7,7 +7,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/memtable_allocator.h"
+#include "memtable/memtable_allocator.h"
 
 #include <assert.h>
 #include "rocksdb/write_buffer_manager.h"
