+++ memtable/skiplist_test.cc
@@ -7,7 +7,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/skiplist.h"
+#include "memtable/skiplist.h"
 #include <set>
 #include "rocksdb/env.h"
 #include "util/arena.h"
