+++ utilities/write_batch_with_index/write_batch_with_index.cc
@@ -14,7 +14,7 @@
 #include "db/db_impl.h"
 #include "db/merge_context.h"
 #include "db/merge_helper.h"
-#include "db/skiplist.h"
+#include "memtable/skiplist.h"
 #include "options/db_options.h"
 #include "rocksdb/comparator.h"
 #include "rocksdb/iterator.h"
